refactor(generator): split orchestration into analyzer and validator components
- Extracted GeneratorAnalyzer (facts-only analysis pipeline)
- Extracted GeneratorValidator (validation chain + cached validators)
- Simplified Generator to orchestration-only responsibilities
- Improved cohesion and separation of concerns
- No behavioral changes
- All tests remain green

This prepares the generator for final file-by-file polishing
ahead of v1.0 stabilization.
